requirements.txt
absl-py==2.0.0
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
deepface==0.0.79
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
gdown==4.7.1
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.14.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mtcnn==0.1.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scipy==1.11.3
six==1.16.0
soupsieve==2.5
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
wrapt==1.14.1